requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
urllib3==1.26.14
websocket-client==1.5.1
